refactor: separate chain integrations
