+++ tsconfig.json
@@ -9,6 +9,10 @@
 		"outDir": "dist",
 		"strict": true
 	},
-	"include": ["./src/*"],
+	"include": [
+		"./src/*",
+		"./src/unique-chains/*",
+		"./src/acala-chains/*",
+	],
 	"lib": ["es2017"],
 }
